+++ packages/lila-webpack-config/src/index.js
@@ -1,3 +1,4 @@
+import merge from 'webpack-merge';
 import dev from './dev';
 import analyze from './analyze';
 import build from './build';
@@ -7,6 +8,10 @@ export default lila => {
   const { setSetting } = lila;
 
   setSetting('webpackConfigGenerator', webpack => ({ entry, cmd, config }) => {
+    const { extra } = config;
+    const extraWebpackConfig =
+      typeof extra === 'function' ? extra(webpack) : extra;
+
     let webpackConfig = {};
 
     if (cmd === 'dev' || cmd === 'serve')
@@ -23,7 +28,9 @@ export default lila => {
     if (plugins.length && webpackConfig.plugins)
       webpackConfig.plugins.push(...plugins);
 
-    return webpackConfig;
+    return extraWebpackConfig
+      ? merge(webpackConfig, extraWebpackConfig)
+      : webpackConfig;
   });
 
   setSetting('getEntries', makeGetEntries(lila));